core: allow result variadic inference (#3559)
Sort of a revert of #3416, but I believe this fixes the underlying
problem.

Changes range constraint inference so that passing in a length is
optional. The rationale for this is that the length of operands is
always known, but the length of results is never known (as they can
always be omitted). Similarly updates the `can_infer` function.