dialects: (builtin) use truthiness in of constant IntegerAttr values (#3556)
In a PR that depends on this one, I propose to normalize all signless
integers to the signed range. This has an impact on a number of tests
that depend on `false` parsing to an i1 with `int` value of `1`, and not
on the fact that it's not `0`. This PR implements the normalization just
for `i1`, and adjusts the passes that depend on this change.