dialects: (builtin) add support for f16 packing and unpacking (#3605)
The "<e" comes with the following warning:

> The IEEE 754 binary16 “half precision” type was introduced in the 2008
revision of the [IEEE 754
standard](https://en.wikipedia.org/wiki/IEEE_754-2008_revision). It has
a sign bit, a 5-bit exponent and 11-bit precision (with 10 bits
explicitly stored), and can represent numbers between approximately
6.1e-05 and 6.5e+04 at full precision. This type is not widely supported
by C compilers: on a typical machine, an unsigned short can be used for
storage, but not for math operations. See the Wikipedia page on the
[half-precision floating-point
format](https://en.wikipedia.org/wiki/Half-precision_floating-point_format)
for more information.

As we only use it for storage, this should be fine though imo