misc: use .venv by default instead of venv (#3629)
I didn't want to have this discussion before we merged uv, but it
definitely feels like it really prefers `.venv`, and makes lots of
things easier if this name is used. How do we feel about using this name
by default? It can still easily be overridden locally, but will change
to .venv by default for everyone.