transformations: (apply-individual-rewrite) lazily find canonicalization patterns (#3631)
It's time to make this part of the codebase a bit more extensible and
robust, as we will be reusing it for some scheduling exploration. As a
first step, this PR changes the rewrite infrastructure to get the
available patterns on the operation lazily, instead of by scanning all
ops of all dialects.