CI: add lockfile update action to manually trigger for dependabot PRs (#3647)
I tested it [here](https://github.com/xdslproject/xdsl/pull/3646) and it
works if you do the dumb thing and trigger on push, but it seems to not
run actions on unsigned commits by default. I'm hoping that if we
trigger the action manually it'll inherit the clicker's permissions and
re-run the action. In either case, I think whatever we adopt should
supercede the cron job one as we want to never be out of sync between
the pyproject and lockfile.