installation: add mechanism for symlinking mlir-opt when creating venv (#3660)
For the last couple of years, I've had a custom-built llvm bin folder in
my path. This has mostly been OK but sometimes resulted in weird build
errors. With this mechanism, we can have the best of both worlds, with
having a normal mlir-opt in the path except for when the venv is
activated:

```
~/Developer/xdslproject/xdsl sasha/misc/env *129 !1 ❯ which mlir-opt           
/opt/homebrew/opt/llvm/bin/mlir-opt
~/Developer/xdslproject/xdsl sasha/misc/env *129 ❯ uv run which mlir-opt  
/Users/sasha/Developer/xdslproject/xdsl/.venv/bin/mlir-opt
```