transformations: (memref-stream-interleave) always take bigger factor (#3721)
When I initially implemented this pass, I assumed that it was better to
err on the side of caution, to take the smallest factor above 3. In
practice, this isn't the best approach, as a factor of 4 is only good in
the scenario with no function cache misses, and it's a safer bet to take
bigger values if possible. We also learned from the Snitch backend paper
experiments that the register pressure never goes close to the limit for
the kernels we care about, so this should not be that risky.

I'm currently working on automatically pruning the space of possible
values, and merging this PR would help keep the logic relatively simple,
while sharing some helpers between this heurist-based pass, and the new
compile-and-evaluate-based-pass.