interactive: store passes in app instead of class+spec tuples (#3723)
This part of the code was written when passes were not yet immutable, so
it was important to be able to reinitialize them each time. Now that
they are immutable, we can store them directly and avoid the
initialization work.

I didn't want to migrate everything in one go, but the tree is still in
terms of class + spec pairs, once this PR is merged I can open the PR
migrating that part of the codebase.