documentation: (marimo) move pipeline_html helper to utils (#3719)
Creates a unified way of presenting the intermediate stages of
compilation in the marimo notebooks, together with optional descriptions
of what's happening. I'm not super happy with the way it's presented
right now, but we can iterate on it once it's in, and all the notebooks
will benefit from it at the same time. I'm relatively sure that this API
is all we need, but we can change it later also.