documentation: add mkdocs-based documentation (#3754)
It is time to add a docs website to xDSL. I'd like for our docs to be as
pretty as [marimo's](https://docs.marimo.io/api/), with similar
interactive elements and pretty styling. I'm not super confident in
web-things, so would love some help with styling and deploying to either
docs.xdsl.dev or xdsl.dev/docs using GH pages.

This PR adds mkdocs, adds some make commands to build and serve the docs
locally, and adds some barebones docs for the core.py file in xDSL.
In the future, it would be nice to do something like MLIR's automatic
building of docs for dialects and other parts of the codebase that are
more structured.