parser: parse bytes directly into a BytesAttr for DenseIntOrFPElementsAttrs (#3845)
Because DenseIntOrFPElementsAttrs are backed by `bytes`, there is no
need to convert the hex representation to a list of values when creating
the attribute. We can directly convert the hex representation to a
BytesAttr when constructing the attribute.

This is a lot faster, and gets rid of a lot of code somewhat duplicating
the `StructPackable` things we made.