misc: reduce max-line-length ruff setting to 130 [NFC] (#3869)
I don't know if we'll ever get to 88, but it feels like it would be good
to get the line lengths a little more under control. In this PR, I
update the line length, ignore all test files, ignore the line length
limit on all assembly format strings, reflow some doc strings, and split
up some inline strings in the code.

After this change, there are 840 errors when I set the line length to
88, and 313 if I set it to 100. I think without some sort of AI "fix
everything" button it's a waste of time to go shorter, but our current
setting of 300 definitely seems to big.

This PR contains about 60 whitespace changes.