installation: specify find (#3874)
Reopens #3848, which fixes a regression introduced by #3815, and
partially fixed with #3847.

While the issue is fixed on most installations with the other fix PR it
breaks on some systems, and when a `site` folder is present after
generating the docs website locally.

We tested this change with @georgebisbas today, on whose linux machine
xDSL cannot currently be installed, and works after applying the changes
in this PR.