wip wip wip one test not passing
mlir-opt does not print default values of properties in the property dict, but we do, and expect them to be present
type default value of memref not printing correctly, dense<0> should have type tensor<1xindex> not memref<1xindex> (according to mlir-opt)

I wonder if there's a helper missing to print the attr-dict of an op including the properties, without the reserved names. There is such funcitonality in print-op-generic but it's not properly exposed.