bug: (x86) fix the lowering of the function signature (#3938)
This PR concerns the registers used to pass function parameters. The
previous PR retrieved them using get_register, so the function type was
inherently empty (I thought a x86 function was just a label). This new
PR provides the registers in question as well as the stack pointer as
input to the first basic block.

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: Sasha Lopoukhine <superlopuh@gmail.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: compor <compor@users.noreply.github.com>
Co-authored-by: superlopuh <superlopuh@users.noreply.github.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: jumerckx <31353884+jumerckx@users.noreply.github.com>