core: reuse field definition functions across op definitions (#4016)
We can reuse the function instances, so there's no need to redefine them
for every operation definition.

Slight perf improvement for `make filecheck`, but it's annoyingly close
to the margin off error, I look forward to seeing the perf results
later. I also have a feeling that it should help pypy's tracing to have
only one of each of these functions for all of xDSL, instead of one per
op.