documentation: remove database example notebook (#4013)
I just re-read it, and it's so far gone from being correct that it's
kind of impressive. I accept all responsibility for making changes to
the API that have resulted in the final module not making sense.

It mentions things that we talk about in our existing marimo notebooks,
so I don't think we need to move any of the contents.