dialects: (arith) fix `arith.constant` construction (#3916)
I bumped into this recently, and it is documented in #2390, but it is
not the parsing and printing that is broken.
We just use the `value_type` when provided, but don't adjust the type of
the value attribute to it, ending up in having

```mlir
%0 = "arith.constant"() <{value = 0 : i32}> : () -> i64
```

which is invalid IR.

~I'd expect the `TypedAttributeConstraint` to kick in, but even though I
tried to play with it, I'm not sure if I'm missing something obvious
there. Happy to close this if I did.~