backend: use negative indices to track infinite registers (#3964)
The overall goal is to be able to track registers when allocating by
their indices, instead of an instance. On some platforms, 32-bit
registers and 64-bit registers overlap, so the type of the register
class does not uniquely identify the space of non-aliasing registers.
Even in RISC-V, we already have this problem, as the `x` register names
alias with the `a`, `t`, `s`, etc register names.

This PR is a first step in that direction, which lets us uniquely
specify an index for the "infinite" register set for each class.
Subsequent PRs will actually leverage this index.