testing: don't cd to git root dir in lit (#3953)
This change in configuration was motivated by the lack of detection of
missing `__init__.py` files in the project. We now have a ruff check for
this (#3945), but in my changing the config, I found a number of other
issues stemming from the `cd ../..` in the preamble to the tests. One is
that for some reason files with no RUN lines aren't detected with the
preamble. I fixed the broken tests that weren't run in a separate PR
(#3944), so this is just about making the config change and making minor
fixes to remaining files.