core: add custom repr to BaseAttr and ParamAttrConstraint (#4094)
For use in dialect and dialect stub generation. Before this change, the
default repr prints the class as `<class
xdsl.dialects.builtin.IntegerAttr>` or whatever, and now we just print
`IntegerAttr`, which won't work for `eval` if you don't already have the
class in your context, but will work for our use-cases for now.