documentation: embed marimo notebooks in docs website (#3778)
To test locally:

```
make venv
make docs-serve
```

See it live: https://xdsl--3778.org.readthedocs.build/3778/marimo/

I had to add a temporary mkdocs extension to inline the marimo thing,
due to the upstream one being broken. They said that they were
interested in fixing it, but I actually had to reimplement a part of a
compression library that they depend on in js to re-create the URLs, so
I don't want to be blocked on the Marimo team approving that.

Resolves #3852

---------

Co-authored-by: Chris Vasiladiotis <cvassiladiotis@gmail.com>