tool: (xdsl-tblgen) fix assembly format (#4158)
Fixes assembly format printing for xdsl-tblgen by wrapping the string in
triple quotes. This is necessary to deal with newlines, an alternative
would be to replace all the new lines in the assembly format with
spaces, or check if the assembly format has new lines before deciding
whether to use single or triple quotes.