dialects (arm): Add NEON FMUL op (mixed scalar/vector) (#4122)
Add Floating-point multiply operation to ARM (NEON) dialect (mixed:
first source operand is a vector, second is a scalar, destination is a
vector).

This instruction multiplies each of the floating-point values in the
first source operand by the second source operand and writes the
resulting values to the corresponding lanes of the destination.

---------

Co-authored-by: Chris Vasiladiotis <cvassiladiotis@gmail.com>