transformations: no assembly section when lowering to riscv_func (#4161)
Now that the riscv_func.funcs can emit their own .text section as
necessary, we don't need the explicit assembly_section structure. With
this change, we can look up the riscv func using the standard
SymbolOpInterface, like func.funv.

CC @jumerckx - With this the functions will be ported, but I need to do
a similar refactor for global data for the switch to symbols in the
interpreter to work.