dialects: (pdl_interp) Add operations (#4176)
The [pdl_interp
dialect](https://mlir.llvm.org/docs/Dialects/PDLInterpOps/) provides a
lower-level abstraction compared to pdl for executing pattern matching
and rewriting. Lowering from pdl to pdl_interp is not considered in this
pr.
Including `assembly_format` for the operations where this is supported,
and using the generic form in tests for the implemented operations where
the `assembly_format` does not work.

Only a subset of operations is implemented for now.

---------

Co-authored-by: Sasha Lopoukhine <superlopuh@gmail.com>