core: Lazily include ImplicitBuilder in Operation __post_init__ (#4163)
Refactor `ImplicitBuilder` to lazily add its modified `__post_init__`
only when it is in use within a context.

The `ImplicitBuilder` implementation can be further improved by storing
the stack as a linked list in object properties instead of using the
singleton `_ImplicitBuilderStack`, but this is a seperate issue to the
lazy enablement of `ImplicitBuilder`s on operations.

---------

Co-authored-by: Chris Vasiladiotis <cvassiladiotis@gmail.com>