dialects: (x86) Change RR_MovOp and RI_MovOp to have only one input
Before this change, RR_MovOp and RI_MovOp were subclasses of R_RR_Operation and R_RI_Operation, respectively. These classes are meant for instructions for which, in x86 Intel syntax, the first operand refers to the register which is both the first input register and the output register, and the second operand refers to the second input register or immediate.

The mov instruction with 32-bit or 64-bit registers has only one input because the result register is fully overwritten (in the 32-bit case, the upper half is zeroed). In these cases, there is only one input. With 8-bit or 16-bit registers, only part of the result register is overwritten, so the content of the result register depends on the previous content of it. Partial registers except 32-bit ones are not yet implemented in the x86 dialect, so for now the latter case can be ignored. Therefore, the mov instruction is changed to be modelled with one input register or immediate, and one result register. This change is in line with #3967. There are no existing abstract classes for these kinds of instructions, so RR_MovOp and RI_MovOp were changed to directly subclass X86Instruction.

If the result register is a 64-bit register, R_RI_Operation should be able to take 64-bit immediates. However, supporting this is currently hard because unallocated registers don’t track the size of the register. As R_RI_Operation didn’t support 64-bit immediates, either, this is not a regression. (As a side-note, mov is the only x86 instruction able to take 64-bit immediates.)