dialects: (builtin) add generic parameter to DenseIntOrFPElementsAttr (#4292)
A much simpler version of #3492 using the default valued typevars. The
advantage of this is that you can now represent for example
`DenseIntOrFPElementsAttr[AnyFloat]` as a type.

This PR does not attempt to specialise the use of
`DenseIntOrFPElementsAttr` anywhere. I also didn't touch the builder
methods `tensor_from_list` or `vector_from_list` or `from_list` as I
personally believe it might be possible/best to remove the methods in
the future, favouring `create_dense_{float/int/index}`.