misc: (eqsat) eqsat-serialize-egraph pass (#4070)
This can be used to dump a json representation of a piece of IR using
the eqsat dialect. The json can then be used to visualize the e-graph
here https://egraphs-good.github.io/egraph-visualizer/

Serialisation is currently implemented as a pass but should perhaps
instead be a target? Currently, running the pass dumps json as well as
the original IR, meaning that you can't just pipe the output somewhere
directly.

---------

Co-authored-by: Sasha Lopoukhine <superlopuh@gmail.com>