core: (constraints) remove variance on `get_variable_extractors` (#4322)
I'm pretty sure there was some type unsoundness created by having the
parameter on `GenericAttrConstraint` be covariant and then having the
extractors use this type parameter. This was highlighted by #4238 ,
where the `TypedAttributeConstraint` extractor encounters an attribute
which is not a `TypedAttribute`.

This PR refactors `get_variable_extractors` to remove the variant
parameter, adding in some new necessary checks within the extractor.