backend: RegisterQueue -> RegisterStack (#4466)
I was recently made aware of the fact that "Queue" isn't a generic term
for a collection to which elements are added and removed in a few
possible orders, so I'd like to rename this LIFO thing to Stack. This PR
also fuses the concrete implementation with the abstract superclass, as
the abstraction isn't really leveraged anwyhere so isn't really
productive.