core: (Printer) remove variadic prints (#4543)
Removes all instances of using the variadic behaviour of
`Printer.print`. All tests pass if `assert len(argv) == 1` is added to
the top of this function.

I haven't added any sort of deprecation/change to `Printer.print`
because I wasn't sure how and it didn't seem worth it if we were just
going to deprecate the whole thing anyway.

---------

Co-authored-by: Sasha Lopoukhine <superlopuh@gmail.com>