dialects: (riscv) remove uses of Printer.print (#4567)
This is all the cases for riscv. 

I'm unhappy with the `printer.print_string(str(an_integer))` and in
these cases where we have an IntegerAttr we can actually do:
```
integer_attr.print_without_type(printer)
```
but I think this would print `true` and `false` in `i1` cases which may
not be what you always want? In any case what is here preserves the
current behaviour, but I just wanted to point out that I'm uneasy about
printing integers.