core: (Printer) add print_int (#4590)
Series of PRs to try to improve standardisation of print_int and
print_float

Adds a `print_int` function similar to `print_float`.

This deals with some tricky corners, such as `i1` being printed as
`true/false` and that python booleans are apparently also python
integers for function calls. This replaces the need for
`type.print_without_type` that was on `IntegerType` and `IndexType`.

Also fixes a bug that `si1` and `ui1` were being printed as `true/false`
whereas in mlir this is only true for `i1` (as far as I can observe
anyway)