dialects: (llvm) Fixed bug in GlobalOp where body was always empty when initialised (#4689)
The llvm.GlobalOp operation can have a body associated with it, however
the operation's initialisation in xDSL always set this to be empty.
Therefore have added an optional body parameter to the function so that
this can be set by the caller.

The xDSL MLIR filecheck test for globalop already has a body associated
with it

---------

Co-authored-by: Alex Rice <alexrice999@hotmail.co.uk>