dialects: (math) Added fastmath property to fp operations in math that were missing it (#4693)
Several operations in the math dialect were missing the fastmath
property, this was likely because in a previous version of MLIR that was
correct but now all FP operations include the fastmath flag. Therefore
have removed the distinction between the base unary/binary FP operation
without, and the one with, fastmath, and all FP operations include
fastmath now as a property.

Added MLIR based tests with fastmath flag in the IR

---------

Co-authored-by: Sasha Lopoukhine <superlopuh@gmail.com>