core: (operation) Added operation name to error raised when verifying properties and attributes (#4696)
Verification for properties and attributes on an operation raises an
exception when trying to set a property/attribute that doesn't exist,
however it just reports the property/attribute name that is trying to be
set and not the operation that is being acted upon. This makes it quite
difficult to track down issues in a large IR (e.g. when parsing).

This PR adds the operation name to the error message, making it easier
to debug issues