dialects:(omp) add `omp.target_data` op (#4706)
This also makes `map_type` and `map_capture_type` required properties of
`omp.map.info`. Upstream, in `llvmorg-20.1.7` they are defined as
`OptionalAttr<UI64Attr>` and `OptionalAttr<VariableCaptureKindAttr>`
respectively, but later assumed to be present, this appears to be a bug.

Both are non-optional on HEAD.

`TargetEnterDataOp`, `TargetExitDataOp` and `TargetUpdateOp` will follow
in a separate PR.