dialects:(omp) add task based data mapping ops
Also adds an extra constraint on mapped values which was missed in the
previous PR: When the `map_type` is constrained to one of a number of
flags, these constraints are on the `var_ptr` operand to `omp.map.info`,
not on the result of `omp.map.info`. So we need to keep track of which
flags have already been used with which operands.

**Operations added:**

* TargetEnterDataOp
* TargetExitDataOp
* TargetUpdateOp