Don't backtrack in rewriter's finalize
Finalize is interpreted by going back to the last backtrackingpoint.
This is correct for `finalize`s that occur in the matching portion of a
pdl_interp file, but isn't for the finalize at the end of an actual
rewrite.
Previously, this lead to very deep recursion, since finalize would not
return from the call but rather jump back to a point in the matcher.
We now keep track of whether the interpreter is in a matching context.