backend:(csl) add extension point for the printer
This extension point would allow someone to subclass `CslPrintContext`
in order to change how existing operations are printed or to add support
for new ones.

Most of the functionality of the printer can be modified by overriding
existing methods, but the logic for printing operations is all in
`print_block` and taking it out would be a much more substantial change.

The sub-classed printer can also not be passed to print_to_csl as an
optional third argument.