core: (SymbolNameConstraint) add assembly format symbol names (#4811)
Adds a new constraint `SymbolNameConstr`, which takes the role that
`SymbolNameAttr` does in MLIR (note that `SymbolNameAttr` is not an
attribute in MLIR, but some tablegen construct that I think translates
to a constraint).
The assembly format then uses this information to parse/print the
attribute as a symbol. Made this work with optional attributes and
optional groups.