dialects: (linalg) Broadcast and transpose op init methods take sequence of result types (#4892)
This is a bug fix for the broadcast and transpose linalg methods, where
the result type is varadic but the default result init method is None,
which results in an error. This therefore fixes that by defaulting to an
empty list.

This also ensures these two operations match MLIR behaviour which is
when the outs type is a tensor then the result type of the operation is
a tensor, otherwise if the outs type is a memref then there is no result
type. The parsing routing did not handle this, so this nuance is now
added (the verification already picked this up, as the only result type
can be a tensor). This meant we could not use memrefs as output from
these operations when passed to MLIR, which we can now.

I have updated the filecheck test to include no result type with memref
as outs type