dialects: (vector) use custom assembly for vector.extract (#4883)
Example use of custom directives.

I believe the change is spacing is actually a bugfix, as it agrees with
upstream documentation, but I could likely split out this change or
revert it if wanted.

This has strengthened my belief that it would be a good idea to define
some getters and setters on the common format directives.