dialects: (hlfir) fixed miss-matched order of operands (#4936)
There were a few operations in HLFIR where the operands were in the
wrong order, this is because HLFIR tends to mix them around between
operations. This is fixed, also improved code commenting and also
alphabetical ordering of operands in dialect definition to make easier
to identify what is present.

This is an experimental dialect