dialects: (dlti) Add initialisation convenience from dictionary and python types (#4952)
DLTI dialect initialises the entry map operations based on a dictionary
or list, and the entries themselves are initialised based on str or
attribute (for key), and attribute, str, int or float for value. This is
for convenience when creating these in code.

Pyright test added for this, along with testing verification of aspects
in the dialect too