installation: try renovate nix support (#5021)
I guess we try the default config first and see what happens. From
reading the docs this should try to update the lock file once a week.
I'm not 100% sure if this is a great idea without any CI but I guess I
can manually test the commits for now.