core: (Operation) simplify `irdl_op_verify_arg_list` (#5022)
Step one of simplifying some of the variadics infrastructure. This
removes the last use of the return value for `get_variadic_sizes`, which
should allow that method to be simplified. We instead rely on the
operation accessors to get the various variadic ranges.

I've kept the logic to put the slice in the error message instead of
referring to it by construct name, as I fear the resulting diff from
changing all the tests would be very large.

---------

Co-authored-by: Sasha Lopoukhine <superlopuh@gmail.com>