documentation: add test page for mkdocs-marimo (#5062)
This PR adds a simple documentation test of `mkdocs-marimo`, which reads
a string from a text box and prints it again. This allows us to test
mkdocs-marimo's update to offer real-time reactive updates.

We currently set `debounce = False` as the notebook does not work
reliably if this is set to an integer debounce value. However, we now
get per-character updates as one types.

We also make the flickering spinner invisible.

This PR was extracted from
https://github.com/xdslproject/xdsl/pull/5060.

---------

Co-authored-by: Sasha Lopoukhine <superlopuh@gmail.com>