documentation: make mkdocs-marimo work on 'latest' branch
When the documentation is shipped to readthedocs.io, the wheel is either
in the `latest/` or `stable/` branch, but it is currently unclear how
to distinguish. As we currently look in `/` which is universally wrong,
I change this to `latest/` hoping to fing a more principled soluation
over the next weeks.