documentation: marimo expressions - make pass pipeline work (#5124)
This adds a basic pass pipeline, which is not yet hooked up to an
interactive flow. However, we add a single slider. It seems currently
this slider does not work as the CSS information needed to display it is
unavailable. We push this anyhow, to have an online instance where this
can be debugged.

This has been extracted from
https://github.com/xdslproject/xdsl/pull/5122.