dialects: (builtin) Update `DYNAMIC_INDEX` to match C++ MLIR (#5273)
In C++ MLIR dynamic indices are specified by kDynamicShape which
represents the smallest integer in `i64`: `-(2 ** 63)`. This PR updates
xDSL to use this value as it is required for printing some ops in
generic format, e.g.:

```
mesh.mesh @mesh0(shape = ?x2)
```

ran with `mlir-opt --mlir-print-op-generic` becomes:

```
"builtin.module"() ({
  "mesh.mesh"() <{shape = array<i64: -9223372036854775808, 2>, sym_name = "mesh0"}> : () -> ()
}) : () -> ()
```

Meaning this is required for generic MLIR compatibility. The previous
code often used `-1` as a value instead of `DYNAMIC_INDEX` so this also
updates that, but it's possible that some untested code isn't updated in
this PR.

---------

Co-authored-by: Sasha Lopoukhine <superlopuh@gmail.com>