transformations: (apply-eqsat-pdl-interp) Support e-class analyses (#5407)
EqsatPDLInterpFunctions now carries a list of analyses, currently
required to be of type `SparseForwardDataFlowAnalysis[Lattice[Any]]`.

When a new operation is inserted (`run_create_operation`), the
`visit_operation_impl` of the analysis is run to get the initial
analysis value (also for the new e-class encompassing the operation).

When e-classes are merged (`run_replace`), their analysis data is joined
with the lattice's `meet` method.

Lastly, invariance maintenance in `repair` follows the description of
e-class analyses in egg. Because `visit_operation_impl` `join`s the
state that it computes with the existing state, we first set the
existing state to uninitialized (`result._value = result.value_cls()`)
such that the join operation with the new state simply yields that new
state.