core: (irdl) Make irdl_options immutable (#5511)
Makes `irdl_options` immutable by enforcing that they should be a
`tuple[IRDLOption, ...]` instead of a `list[IRDLOption]`.
Maintains backwards compatibility with a deprecation warning that can be
removed at some convenient point in the future.

This change helps to ensure down-stream projects with more strict
linting requirements don't run into issues with their operations
defining a class variable `irdl_options` that is mutable but not
annotated as a `ClassVar[...]` (since there are other rules for
`ClassVar`s that make it incompatible for IRDL).